brightness should be an integer

+++ src/Engine/Core/State.hs
@@ -40,7 +40,7 @@ data EngineEnv = EngineEnv
   , windowStateRef     ∷ IORef WindowState
   , framebufferSizeRef ∷ IORef (Int, Int)
   , fpsRef             ∷ IORef Double
-  , brightnessRef      ∷ IORef Float
+  , brightnessRef      ∷ IORef Int
   , eventQueue         ∷ Q.Queue Event
   , inputQueue         ∷ Q.Queue InputEvent
   , loggerRef          ∷ IORef LoggerState
